Skip nonexistent repos when removing a grade - close #17

+++ src/utils/grades/remove/remove.ts
@@ -1,11 +1,12 @@
 import ora from 'ora'
 import slug from 'slug'
+import chalk from 'chalk'
 
 import { Classroom } from '@/api/classroom.js'
 import { getConfigRepo, updateConfigRepo } from '@/utils/config/repo/repo.js'
 import { Assignments } from '@/api/assignment.js'
-
 import { deleteRepo } from '@/api/repo.js'
+import { repoExists } from '@/utils/repos/repos.js'
 
 export default async function remove(name: string, assignment: Assignments[number], classroom: Classroom) {
     const spinner = ora(`Removing grade ${name} from ${assignment.title}`).start()
@@ -25,14 +26,32 @@ export default async function remove(name: string, assignment: Assignments[numbe
     const repoPrefix = `${assignment.slug}-${slug(name)}-`
 
     spinner.text = 'Deleting teacher repository'
-    await deleteRepo(org, repoPrefix + 'teacher')
+    if (await repoExists(org, repoPrefix + 'teacher')) {
+        await deleteRepo(org, repoPrefix + 'teacher')
+    } else {
+        spinner.stop()
+        console.log(chalk.yellow('Teacher repository does not exist'))
+        spinner.start()
+    }
 
     spinner.text = 'Deleting grader repository'
-    await deleteRepo(org, repoPrefix + 'grader')
+    if (await repoExists(org, repoPrefix + 'grader')) {
+        await deleteRepo(org, repoPrefix + 'grader')
+    } else {
+        spinner.stop()
+        console.log(chalk.yellow('Grader repository does not exist'))
+        spinner.start()
+    }
 
     for (const anonymousNameMap of grade.anonymousNamesMap) {
         spinner.text = `Deleting ${anonymousNameMap.studentName}'s anonymous repository`
-        await deleteRepo(org, repoPrefix + anonymousNameMap.anonymousName)
+        if (await repoExists(org, repoPrefix + anonymousNameMap.anonymousName)) {
+            await deleteRepo(org, repoPrefix + anonymousNameMap.anonymousName)
+        } else {
+            spinner.stop()
+            console.log(chalk.yellow(`Repository for ${anonymousNameMap.anonymousName} does not exist`))
+            spinner.start()
+        }
     }
 
     configRepo.grades.splice(gradeExistsIndex, 1)
